app/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sse-starlette
databricks-sdk
python-dotenv
